Refactor displayAllBooks into module
